test(lyrics): add merge_chunk_lines unit coverage + split tests to sibling
#65 asked for unit coverage on align_chunked. The function itself can't
be unit-tested without standing up real ffmpeg and a Replicate mock —
both deferred to integration tests. The highest-value invariant inside
it (chunk-overlap dedup + global-offset translation) is now extracted
as the pure helper `merge_chunk_lines`, and gets 6 unit tests covering:

- First chunk (idx=0) passes everything through with offset 0
- Subsequent chunks translate chunk-local → global ms
- Overlap-region lines dropped (chunk N's first overlap is chunk N-1's tail)
- Word-level timings get the same offset applied
- Boundary case: line at exact drop_below_ms is KEPT (strict <)
- Empty chunk yields empty

`align_chunked` keeps its `mutants::skip` justification (ffmpeg + HTTP
required); merge_chunk_lines does NOT skip — the pure path is what
the new tests exercise.

Tests outgrew the parent file (1088 lines, over the 1000 cap) so the
existing tests module is split into the sibling
`whisperx_replicate_tests.rs` via #[path], mirroring #95/#104.

whisperx_replicate.rs is now 392 lines; whisperx_replicate_tests.rs is 700.

Closes #65
